+++ nl/hanze/projectOOP/mvc/Main.java
@@ -1,4 +1,4 @@
-package Parkeersimulator;
+package nl.hanze.projectOOP.mvc;
 public class Main
 {
     public static void main(String[] args)
