+++ nl/hanze/projectOOP/mvc/model/AdHocCar.java
@@ -1,16 +1,15 @@
-package Parkeersimulator;
-
+package nl.hanze.projectOOP.mvc;
 import java.util.Random;
 import java.awt.*;
 
-public class ParkingPassCar extends Car {
-	private static final Color COLOR=Color.blue;
+public class AdHocCar extends Car {
+	private static final Color COLOR=Color.red;
 	
-    public ParkingPassCar() {
+    public AdHocCar() {
     	Random random = new Random();
     	int stayMinutes = (int) (15 + random.nextFloat() * 3 * 60);
         this.setMinutesLeft(stayMinutes);
-        this.setHasToPay(false);
+        this.setHasToPay(true);
     }
     
     public Color getColor(){
