+++ nl/hanze/projectOOP/mvc/model/Car.java
@@ -1,5 +1,4 @@
-package Parkeersimulator;
-
+package nl.hanze.projectOOP.mvc;
 import java.awt.*;
 
 public abstract class Car {
