+++ nl/hanze/projectOOP/mvc/model/ParkingPassCar.java
@@ -1,16 +1,15 @@
-package Parkeersimulator;
-
+package nl.hanze.projectOOP.mvc;
 import java.util.Random;
 import java.awt.*;
 
-public class AdHocCar extends Car {
-	private static final Color COLOR=Color.red;
+public class ParkingPassCar extends Car {
+	private static final Color COLOR=Color.blue;
 	
-    public AdHocCar() {
+    public ParkingPassCar() {
     	Random random = new Random();
     	int stayMinutes = (int) (15 + random.nextFloat() * 3 * 60);
         this.setMinutesLeft(stayMinutes);
-        this.setHasToPay(true);
+        this.setHasToPay(false);
     }
     
     public Color getColor(){
