+++ nl/hanze/projectOOP/mvc/view/Graph.java
@@ -1,4 +1,4 @@
-package Parkeersimulator;
+package nl.hanze.projectOOP.mvc;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Graphics;
